Refactor DevTools sidebar code.
