+++ packages/devtools_app/lib/src/standalone_ui/vs_code/flutter_panel.dart
@@ -15,7 +15,7 @@ import '../api/dart_tooling_api.dart';
 import '../api/impl/dart_tooling_api.dart';
 import 'debug_sessions.dart';
 import 'devices.dart';
-import 'devtools.dart';
+import 'devtools/devtools.dart';
 
 /// A general Flutter sidebar panel for embedding inside IDEs.
 ///
